requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
Pygments==2.18.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
